Script fixes

+++ files/root/.bash_aliases
@@ -10,13 +10,13 @@ los()
 		echo $dest
 		if [[ ! "$(basename $1)" =~ "bpiwrt_v" ]]; then
 			for part in ${dev}p*; do
-				sudo mkdir -p ${dest} && sudo mount ${part} ${dest} || return
+				sudo mkdir -p ${dest} && sudo mount ${part} ${dest}
 			done
 		else
 			sudo mkdir -p ${dest}
-			sudo mount ${dev}p2 ${dest} || return
+			sudo mount ${dev}p2 ${dest}
 			sudo mkdir ${dest}/boot 2> /dev/null
-			sudo mount ${dev}p1 ${dest}/boot || return
+			sudo mount ${dev}p1 ${dest}/boot
 		fi
 	fi
 }
@@ -26,16 +26,24 @@ losd()
 		echo "Syntax: losd [loop device number]"
 		return 0
 	fi
-	dev=/dev/loop${1}
-	if ! losetup -l | grep "${dev}" >& /dev/null; then
-		dev=$(losetup -l | grep "$1" | cut -d" " -f 1)
+	[[ "$1" =~ "{$1/mnt/dev}" ]] && dev=${1/mnt/dev}
+	if [[ -f /dev/loop${1} ]]; then
+		dev=/dev/loop${1}
+	else
+		[[ ! -z "$1" ]] && dev=$(losetup -l | grep "${1}" | cut -d" " -f 1)
+		if [[ -z "${dev}" ]]; then
+			dev=$(mount | grep "${1} " | head -1 | awk '{print $1}')
+			dev=${dev/p2/}
+		fi
 	fi
-	if ! losetup -l | grep "${dev}" >& /dev/null; then
+	if [[ -z "${dev}" ]]; then
+		echo "ERROR: No image mounted for \"${1}\".  Aborting..."
+	elif ! losetup -l | grep "${dev}" >& /dev/null; then
 		echo "ERROR: No image mounted for \"${1}\".  Aborting..."
 	else
 		for part in $(mount | grep ${dev}p | awk '{print $3}' | sort -r); do 
-			sudo umount $part || return
-			sudo rmdir $part
+			sudo umount $part
+			[[ ! "${part}" =~ /boot$ ]] && sudo rmdir $part 2> /dev/null
 		done
 		sudo losetup -d "$dev"
 	fi
